requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap5==2024.2
diff-match-patch==20230430
dj-static==0.0.6
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-crispy-forms==2.2
django-import-export==4.1.0
django-js-asset==2.2.0
pillow==10.4.0
reportlab==4.2.2
sqlparse==0.5.0
static-ranges==0.2.0
static3==0.7.0
tablib==3.5.0
tzdata==2024.1
